requirements.txt
requests~=2.*
python-dotenv~=1.0.*
cryptography~=44.0.*
